Upgrade to .net 6
