+++ global.json
@@ -1,6 +1,7 @@
 {
     "sdk": {
-        "version": "5.0.100",
-        "rollForward": "latestFeature"
+        "version": "6.0.300",
+        "rollForward": "latestFeature",
+        "allowPrerelease": false
     }
 }
